change typings
