+++ src/bp/common/typings.ts
@@ -1,4 +1,12 @@
-import { BotDetails, BoxedVarConstructable, Flow, FlowNode, RolloutStrategy, StageRequestApprovers } from 'botpress/sdk'
+import {
+  BotDetails,
+  BoxedVarConstructable,
+  Flow,
+  FlowNode,
+  PromptConstructable,
+  RolloutStrategy,
+  StageRequestApprovers
+} from 'botpress/sdk'
 import { StrategyUser } from 'botpress/sdk'
 import { Request } from 'express'
 
@@ -301,6 +309,29 @@ export interface FormField {
     contextMenu?: FormContextMenu[] // you can add a contextual menu to add extra options
   }
 }
+
+export interface PromptDefinition {
+  id: string
+  config: PromptConfig
+  prompt: PromptConstructable
+}
+
+/** The configuration of the prompt which is saved on the flow */
+export type PromptConfig = {
+  /** An ID used internally to refer to this prompt */
+  type: string
+  /** The label displayed in the studio */
+  label?: string
+  icon?: string
+  /** The ID representing the type of value that is collected by this prompt */
+  valueType?: string
+  /** A list of ID represented by the type of values collected by this prompt */
+  valueTypes?: string[]
+  /** The minimum confidence required for the value to be considered valid */
+  minConfidence?: number
+  /** Whatever happens, the prompt will never ask the user to validate the provided value */
+  noValidation?: boolean
+} & FormDefinition
 
 export interface FormDefinition {
   advancedSettings: FormAdvancedSetting[]
