+++ src/bp/sdk/botpress.d.ts
@@ -136,7 +136,7 @@ declare module 'botpress/sdk' {
     translations?: { [lang: string]: object }
     /** List of new conditions that the module can register */
     dialogConditions?: Condition[]
-    prompts?: PromptDefinition[]
+    prompts?: any
     variables?: any
     /** Called once the core is initialized. Usually for middlewares / database init */
     onServerStarted?: (bp: typeof import('botpress/sdk')) => Promise<void>
@@ -1514,31 +1514,6 @@ declare module 'botpress/sdk' {
     asChatUser?: boolean
   }
 
-  export interface PromptDefinition {
-    id: string
-    config: PromptConfig
-    prompt: PromptConstructable<Prompt>
-  }
-
-  /** The configuration of the prompt which is saved on the flow */
-  export interface PromptConfig {
-    /** An ID used internally to refer to this prompt */
-    type: string
-    /** The label displayed in the studio */
-    label?: string
-    icon?: string
-    /** The ID representing the type of value that is collected by this prompt */
-    valueType?: string
-    /** A list of ID represented by the type of values collected by this prompt */
-    valueTypes?: string[]
-    /** List of custom parameters that will be asked by the prompt */
-    params?: { [paramName: string]: ConditionParam }
-    /** The minimum confidence required for the value to be considered valid */
-    minConfidence?: number
-    /** Whatever happens, the prompt will never ask the user to validate the provided value*/
-    noValidation?: boolean
-  }
-
   export interface PromptNode {
     type: string
     /** The name of the variable that will be filled with the value extracted */
@@ -1583,7 +1558,7 @@ declare module 'botpress/sdk' {
     message?: string
   }
 
-  export interface PromptConstructable<T> {
+  export interface PromptConstructable {
     new (ctor: any): Prompt
   }
